requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic>=2.9.0,<3.0.0